+++ allensdk/brain_observatory/behavior/session_apis/data_io/behavior_ophys_lims_api.py
@@ -100,7 +100,7 @@ class BehaviorOphysLimsExtractor(OphysLimsExtractor, BehaviorLimsExtractor,
         return self.lims_db.fetchone(query, strict=True)
 
     @memoize
-    def get_experiment_container_id(self) -> int:
+    def get_ophys_container_id(self) -> int:
         """Get the experiment container id associated with the ophys
         experiment id used to initialize the API"""
         query = """