+++ allensdk/test/brain_observatory/behavior/test_behavior_project_cache.py
@@ -7,24 +7,14 @@ import logging
 
 from allensdk.brain_observatory.behavior.behavior_project_cache \
     import BehaviorProjectCache
+from allensdk.test.brain_observatory.behavior.conftest import get_resources_dir
 
 
 @pytest.fixture
 def session_table():
     return (pd.DataFrame({"behavior_session_id": [3],
-                          "foraging_id": [1],
                           "ophys_experiment_id": [[5, 6]],
-                          "date_of_acquisition": np.datetime64('2020-02-20'),
-                          "reporter_line": ["Ai93(TITL-GCaMP6f)"],
-                          "driver_line": [["aa"]],
-                          'full_genotype': [
-                              'Vip-IRES-Cre/wt;Ai148(TIT2L-GC6f-ICL-tTA2)/wt',
-                              ],
-                          'cre_line': ['Vip-IRES-Cre'],
-                          'session_type': ['OPHYS_1_images_A'],
-                          'mouse_id': [1],
-                          'session_number': [1],
-                          'indicator': ['GCaMP6f']
+                          "date_of_acquisition": np.datetime64('2020-02-20')
                           }, index=pd.Index([1], name='ophys_session_id'))
             )
 
@@ -50,13 +40,7 @@ def behavior_table():
                           'session_type': ['TRAINING_1_gratings',
                                            'TRAINING_1_gratings',
                                            'OPHYS_1_images_A'],
-                          'session_number': [None, None, 1],
-                          'mouse_id': [1, 1, 1],
-                          'prior_exposures_to_session_type': [0, 1, 0],
-                          'prior_exposures_to_image_set': [
-                              np.nan, np.nan, 0],
-                          'prior_exposures_to_omissions': [0, 0, 0],
-                          'indicator': ['GCaMP6f', 'GCaMP6f', 'GCaMP6f']
+                          'mouse_id': [1, 1, 1]
                           })
             .set_index("behavior_session_id"))
 
@@ -65,30 +49,14 @@ def behavior_table():
 def experiments_table():
     return (pd.DataFrame({"ophys_session_id": [1, 2, 3],
                           "behavior_session_id": [1, 2, 3],
-                          "foraging_id": [1, 2, 3],
                           "ophys_experiment_id": [1, 2, 3],
                           "date_of_acquisition": [
                               np.datetime64('2020-02-20'),
                               np.datetime64('2020-02-21'),
                               np.datetime64('2020-02-22')
                           ],
-                          "reporter_line": ["Ai93(TITL-GCaMP6f)",
-                                            "Ai93(TITL-GCaMP6f)",
-                                            "Ai93(TITL-GCaMP6f)"],
-                          "driver_line": [["aa"], ["aa", "bb"], ["cc"]],
-                          'full_genotype': [
-                              'foo-SlcCre',
-                              'Vip-IRES-Cre/wt;Ai148(TIT2L-GC6f-ICL-tTA2)/wt',
-                              'bar'],
-                          'cre_line': [None, 'Vip-IRES-Cre', None],
-                          'session_type': ['TRAINING_1_gratings',
-                                           'TRAINING_1_gratings',
-                                           'OPHYS_1_images_A'],
-                          'mouse_id': [1, 1, 1],
-                          'session_number': [None, None, 1],
                           'imaging_depth': [75, 75, 75],
-                          'targeted_structure': ['VISp', 'VISp', 'VISp'],
-                          'indicator': ['GCaMP6f', 'GCaMP6f', 'GCaMP6f']
+                          'targeted_structure': ['VISp', 'VISp', 'VISp']
                           })
             .set_index("ophys_experiment_id"))
 
@@ -113,6 +81,7 @@ def mock_api(session_table, behavior_table, experiments_table):
 
         def get_behavior_stage_parameters(self, foraging_ids):
             return {x: {} for x in foraging_ids}
+
     return MockApi
 
 
@@ -134,12 +103,12 @@ def test_get_session_table(TempdirBehaviorCache, session_table):
         path = cache.manifest.path_info.get("ophys_sessions").get("spec")
         assert os.path.exists(path)
 
-    # These get merged in
-    session_table['prior_exposures_to_session_type'] = [0]
-    session_table['prior_exposures_to_image_set'] = [0.0]
-    session_table['prior_exposures_to_omissions'] = [0]
+    expected_path = os.path.join(get_resources_dir(), 'project_metadata',
+                                 'expected')
+    expected = pd.read_pickle(os.path.join(expected_path,
+                                           'ophys_session_table.pkl'))
 
-    pd.testing.assert_frame_equal(session_table, obtained)
+    pd.testing.assert_frame_equal(expected, obtained)
 
 
 @pytest.mark.parametrize("TempdirBehaviorCache", [True, False], indirect=True)
@@ -149,7 +118,13 @@ def test_get_behavior_table(TempdirBehaviorCache, behavior_table):
     if cache.cache:
         path = cache.manifest.path_info.get("behavior_sessions").get("spec")
         assert os.path.exists(path)
-    pd.testing.assert_frame_equal(behavior_table, obtained)
+
+    expected_path = os.path.join(get_resources_dir(), 'project_metadata',
+                                 'expected')
+    expected = pd.read_pickle(os.path.join(expected_path,
+                                           'behavior_session_table.pkl'))
+
+    pd.testing.assert_frame_equal(expected, obtained)
 
 
 @pytest.mark.parametrize("TempdirBehaviorCache", [True, False], indirect=True)
@@ -160,12 +135,12 @@ def test_get_experiments_table(TempdirBehaviorCache, experiments_table):
         path = cache.manifest.path_info.get("ophys_experiments").get("spec")
         assert os.path.exists(path)
 
-    # These get merged in
-    experiments_table['prior_exposures_to_session_type'] = [0, 1, 0]
-    experiments_table['prior_exposures_to_image_set'] = [np.nan, np.nan, 0]
-    experiments_table['prior_exposures_to_omissions'] = [0, 0, 0]
+    expected_path = os.path.join(get_resources_dir(), 'project_metadata',
+                                 'expected')
+    expected = pd.read_pickle(os.path.join(expected_path,
+                                           'ophys_experiment_table.pkl'))
 
-    pd.testing.assert_frame_equal(experiments_table, obtained)
+    pd.testing.assert_frame_equal(expected, obtained)
 
 
 @pytest.mark.parametrize("TempdirBehaviorCache", [True], indirect=True)
@@ -175,16 +150,16 @@ def test_session_table_reads_from_cache(TempdirBehaviorCache, session_table,
     cache = TempdirBehaviorCache
     cache.get_session_table()
     expected_first = [
-        ('call_caching', 20, 'Reading data from cache'),
-        ('call_caching', 20, 'No cache file found.'),
-        ('call_caching', 20, 'Fetching data from remote'),
-        ('call_caching', 20, 'Writing data to cache'),
-        ('call_caching', 20, 'Reading data from cache'),
-        ('call_caching', 20, 'Reading data from cache'),
-        ('call_caching', 20, 'No cache file found.'),
-        ('call_caching', 20, 'Fetching data from remote'),
-        ('call_caching', 20, 'Writing data to cache'),
-        ('call_caching', 20, 'Reading data from cache')]
+        ('call_caching', logging.INFO, 'Reading data from cache'),
+        ('call_caching', logging.INFO, 'No cache file found.'),
+        ('call_caching', logging.INFO, 'Fetching data from remote'),
+        ('call_caching', logging.INFO, 'Writing data to cache'),
+        ('call_caching', logging.INFO, 'Reading data from cache'),
+        ('call_caching', logging.INFO, 'Reading data from cache'),
+        ('call_caching', logging.INFO, 'No cache file found.'),
+        ('call_caching', logging.INFO, 'Fetching data from remote'),
+        ('call_caching', logging.INFO, 'Writing data to cache'),
+        ('call_caching', logging.INFO, 'Reading data from cache')]
     assert expected_first == caplog.record_tuples
     caplog.clear()
     cache.get_session_table()
@@ -198,15 +173,20 @@ def test_behavior_table_reads_from_cache(TempdirBehaviorCache, behavior_table,
     cache = TempdirBehaviorCache
     cache.get_behavior_session_table()
     expected_first = [
-        ("call_caching", logging.INFO, "Reading data from cache"),
-        ("call_caching", logging.INFO, "No cache file found."),
-        ("call_caching", logging.INFO, "Fetching data from remote"),
-        ("call_caching", logging.INFO, "Writing data to cache"),
-        ("call_caching", logging.INFO, "Reading data from cache")]
+        ('call_caching', logging.INFO, 'Reading data from cache'),
+        ('call_caching', logging.INFO, 'No cache file found.'),
+        ('call_caching', logging.INFO, 'Fetching data from remote'),
+        ('call_caching', logging.INFO, 'Writing data to cache'),
+        ('call_caching', logging.INFO, 'Reading data from cache'),
+        ('call_caching', logging.INFO, 'Reading data from cache'),
+        ('call_caching', logging.INFO, 'No cache file found.'),
+        ('call_caching', logging.INFO, 'Fetching data from remote'),
+        ('call_caching', logging.INFO, 'Writing data to cache'),
+        ('call_caching', logging.INFO, 'Reading data from cache')]
     assert expected_first == caplog.record_tuples
     caplog.clear()
     cache.get_behavior_session_table()
-    assert [expected_first[0]] == caplog.record_tuples
+    assert [expected_first[0], expected_first[-1]] == caplog.record_tuples
 
 
 @pytest.mark.parametrize("TempdirBehaviorCache", [True, False], indirect=True)